backend/requirements.txt
# === CORE FRAMEWORK ===
Flask==3.0.0
Flask-CORS==4.0.1
Werkzeug==3.0.1

# === DATABASE + ORM ===
prisma==0.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# === AWS S3 STORAGE ===
boto3==1.34.130

# === PAYMENTS ===
stripe==11.0.0

# === AI MODERATION ===
nudenet==1.2.0
opencv-python-headless==4.10.0.82
numpy==1.26.4
Pillow==10.4.0
exifread==3.0.0

# === UTILITIES ===
requests==2.31.0
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
gunicorn==22.0.0

# === DEVELOPMENT ===
pytest==8.3.2
black==24.8.0
flake8==7.1.0
